+++ CHANGELOG.md
@@ -1,5 +1,29 @@
-## 2.2.7 (unreleased)
+## 2.4.1 (unreleased)
 
+- Added cohorts for MySQL
+
+## 2.4.0 (2020-12-15)
+
+- Added cohorts
+- Fixed broken routes for some applications
+- Forecasting and uploads are no longer experimental
+
+## 2.3.1 (2020-11-23)
+
+- Improved column names for uploads
+
+## 2.3.0 (2020-11-16)
+
+- Added support for archiving queries
+- Added experimental support for uploads
+
+## 2.2.8 (2020-11-01)
+
+- Fixed error when deleting dashboard
+
+## 2.2.7 (2020-09-07)
+
+- Use `datetime` type in migration
 - Fixed unpermitted parameters on dashboard page
 - Fixed deprecation warnings in Ruby 2.7